refactor: Split rules into search part and rules part
