+++ src/ducks/recurrence/DebugRecurrencePage.jsx
@@ -1,33 +1,29 @@
 import React, { useState, useMemo, useCallback } from 'react'
-import { useQuery } from 'cozy-client'
-import { transactionsConn } from 'doctypes'
-import {
-  findRecurringBundles,
-  getRulesFromConfig,
-  rulesPerName,
-  groupBundles,
-  sameFirstLabel,
-  addStats
-} from './rules'
-import { getAutomaticLabelFromBundle } from './utils'
-import Card from 'cozy-ui/transpiled/react/Card'
-import { Caption, SubTitle } from 'cozy-ui/transpiled/react/Text'
-import { Padded } from 'components/Spacing'
+
 import setWith from 'lodash/setWith'
 import sortBy from 'lodash/sortBy'
 import clone from 'lodash/clone'
+import groupBy from 'lodash/groupBy'
+
+import { useQuery, useClient } from 'cozy-client'
+import Card from 'cozy-ui/transpiled/react/Card'
+import Alerter from 'cozy-ui/transpiled/react/Alerter'
+import Button from 'cozy-ui/transpiled/react/Button'
+import { Caption, SubTitle } from 'cozy-ui/transpiled/react/Text'
+
+import Loading from 'components/Loading/Loading'
+import { Padded } from 'components/Spacing'
+import { transactionsConn } from 'doctypes'
 import tree from 'ducks/categories/tree'
+
 import defaultConfig from './config.json'
-import maxBy from 'lodash/maxBy'
-import minBy from 'lodash/minBy'
-import { useClient } from 'cozy-client'
 import { saveBundles, resetBundles } from './api'
 import RulesDetails from './RulesDetails'
 import DateSlider from './DateSlider'
+import { getRulesFromConfig, rulesPerName } from './rules'
+import { getAutomaticLabelFromBundle } from './utils'
+import { findRecurringBundles, updateBundles } from './search'
 import useStickyState from './useStickyState'
-import Alerter from 'cozy-ui/transpiled/react/Alerter'
-import Button from 'cozy-ui/transpiled/react/Button'
-import Loading from 'components/Loading/Loading'
 
 const immutableSet = (object, path, value) => {
   return setWith(clone(object), path, value, clone)
@@ -183,32 +179,6 @@ const RecurrenceBundle = ({ bundle }) => {
       </table>
     </Card>
   )
-}
-
-const ONE_DAY = 86400 * 1000
-
-const updateBundles = (bundles, newTransactions, rules) => {
-  if (!newTransactions.length) {
-    return bundles
-  }
-  const maxDate = new Date(maxBy(newTransactions, 'date').date)
-  const minDate = new Date(minBy(newTransactions, 'date').date)
-  const dateSpan = (maxDate - minDate) / ONE_DAY
-
-  let updatedBundles
-
-  if (dateSpan > 90 && newTransactions.length > 100) {
-    const newBundles = findRecurringBundles(newTransactions, rules)
-    const allBundles = [...bundles, ...newBundles]
-    updatedBundles = groupBundles(allBundles, sameFirstLabel)
-  } else {
-    const newBundles = newTransactions.map(t => ({ ops: [t] }))
-    const allBundles = [...bundles, ...newBundles]
-    updatedBundles = groupBundles(allBundles, sameFirstLabel)
-  }
-
-  updatedBundles = bundles.map(addStats)
-  return updatedBundles
 }
 
 const makeTextFilter = (searchStr, accessor) => {
