+++ src/ducks/recurrence/rules.js
@@ -3,7 +3,6 @@ import sortBy from 'lodash/sortBy'
 import sum from 'lodash/sum'
 import mergeWith from 'lodash/mergeWith'
 import isArray from 'lodash/isArray'
-import unique from 'lodash/uniq'
 import isString from 'lodash/isString'
 import compose from 'lodash/flowRight'
 import some from 'lodash/some'
@@ -59,7 +58,7 @@ const makeStats = operations => {
  * @param  {Array[Function]} predicates
  * @return {Function}
  */
-const overEvery = predicates => item => {
+export const overEvery = predicates => item => {
   for (const predicate of predicates) {
     if (!predicate(item)) {
       return false
@@ -85,68 +84,10 @@ const mergeBundles = bundles => {
   }
 }
 
-const assert = (pred, msg) => {
-  if (!pred) {
-    throw new Error(msg)
-  }
-}
-
 export const groupBundles = (bundles, rules) => {
   const grouper = compose(rules)
   const groups = groupBy(bundles, grouper)
   return Object.values(groups).map(mergeBundles)
-}
-
-/**
- * How rules work:
- *
- * There are different stages that find and process bundles
- *
- * - Some stages filter bundles
- * - Some change the bundles (map)
- * - Some regroup bundles
- *
- * @param  {array} operations
- * @param  {array} rules
- * @return {array} recurring bundles
- */
-export const findRecurringBundles = (operations, rules) => {
-  const groups = groupBy(
-    operations,
-    x => `${x.manualCategoryId || x.automaticCategoryId}/${x.amount}`
-  )
-
-  let bundles = Object.entries(groups).map(([key, ops]) => {
-    const [categoryId, amount] = key.split('/')
-    return {
-      categoryIds: [categoryId],
-      amounts: [parseInt(amount, 10)],
-      key,
-      ops
-    }
-  })
-
-  const groupedByStage = groupBy(rules, rule => rule.stage)
-
-  const stageKeys = Object.keys(groupedByStage).sort()
-  for (let stageKey of stageKeys) {
-    const ruleInfos = groupedByStage[stageKey]
-    assert(
-      unique(ruleInfos.map(r => r.type)).length === 1,
-      'Cannot have multiple types per stage'
-    )
-    const type = ruleInfos[0].type
-    const rules = ruleInfos.map(ruleInfo => ruleInfo.rule)
-    if (type === 'filter') {
-      bundles = bundles.filter(overEvery(rules))
-    } else if (type === 'map') {
-      bundles = bundles.map(compose(rules))
-    } else if (type === 'group') {
-      bundles = groupBundles(bundles, rules)
-    }
-  }
-
-  return bundles
 }
 
 export const sameFirstLabel = bundle => {
